-fixed bug in ancestral states computations that only occured with codon-style partition files of type:  DNA,	 part_1_codon1 = 1-60\3  DNA,	 part_1_codon2 = 2-60\3  DNA,	 part_1_codon3 = 3-60\3
The problem was that pattern compression for ancestral states was disabled to print out the ancestral states in correspondance to the sites
of the input alignment (without pattern compression and re-ordering). The problem is that when we disable pattern compression the code expects
sites belonging to the same model/partition to be stored contiguously in memory which under this codon scheme is not done if site compression is disabled.

In general we need to be very careful with partitioned models with non-contiguous sites and site compression disabled.

Kassian, this may also be relevant to the stuff you wanted to test.

I now integrated site compression into ancestral state computations and then bring the sites back into the original order once they are printed out to file.
I hope that I did not screw up the indices, initial tests suggest that it is correct though. To quote Derrick Zwickl, this is the most difficult part of implementing ancestral states.