-added the possibility to read-in model parameters and a tree with branch lengths  as generated by the -f e (RAxML_binaryModelParamaters.RUN_ID) option into the  Evolutionary placement algorithm via -R
-This will speed up EPA computations in cases were you repeatedly just want to place a
 single sequence into a given reference tree, since the model parameters and branch lengths
 of the reference tree will not need to be re-estimated.

-Current limitations:
 1. does not work with the CAT model
 2. does not work on partitioned datasets
 3. will yield slightly different results than the EPA in standard mode,
    that is without the -R option, but this is actually a thing that can not
    be avoided.

-Example: given a reference data set with 100 taxa 100.phy and a reference tree:
 First optimize model params and branch lengths with -f e:

 ./raxmlHPC-AVX -f e -s 100.phy -m GTRGAMMA -t RAxML_bestTree.T1 -n Evaluate

RAxML will generate a file called: "RAxML_binaryModelParameters.Evaluate"
and a tree file with optimized branch lengths called "RAxML_result.Evaluate"

Then invoke the placement algorithm as follows:

./raxmlHPC-AVX -f v -s 101.phy -m GTRGAMMA -t RAxML_result.Evaluate -R RAxML_binaryModelParameters.Evaluate -n P1

where dataset 101.phy just contains one sequence to be placed via the EPA.

Tests on this single-gene dataset showed that compared to the standard invocation:

./raxmlHPC-AVX -f v -s 101.phy -m GTRGAMMA -t RAxML_result.Evaluate -n P2

This is about 8 times faster in this case.

The actual speedup will depend on the datset size, that is, the larger the reference dataset (100.phy)
the better speedups you will get.